Refactored vispy renderer
